requirements.txt
alembic
cryptography
fastapi
passlib
prometheus_client
PyJWT
PyMySQL
python-dotenv
python-multipart
qdrant-client
redis
requests
sentence-transformers
SQLAlchemy
uvicorn